merge

+++ bench.c
@@ -83,8 +83,8 @@ int main(int argc, char *argv[])
     {
         for (int i = 0; i < loop; i++)
         {
-            int index = rand() % ((int) BUFFER);
-            printf("index: %i", index);
+            int index = (int) rand() % ((int) BUFFER);
+            printf("index: %d", index);
             if (buffer[index] != NULL)
             {
                 dfree(buffer[index]);
